DRYness and cleanup.
