+++ lib/graph.js
@@ -1,14 +1,12 @@
 var Hoek = require('hoek');
 var Joi = require('joi');
-var Formatting = require('./formatting');
+var Utilities = require('./utilities');
 var Schema = require('./schema');
 var Traversal = require('./traversal');
 
 var internals = {};
 
 exports = module.exports = internals.Graph = function (definition) {
-
-    definition = definition || {};
 
     var validated = Joi.validate(definition, Schema.definition);
     Hoek.assert(!validated.error, validated.error);
@@ -353,24 +351,15 @@ internals.Graph.prototype.equals = function (graph, matchWeights) {
         return false;
     }
 
-    var normalizeEdges = function (edgeObj) {
+    var normalizeEdges = function (pairsObj) {
 
         var normalized = {};
 
-        var u;
-        var us = Object.keys(edgeObj);
-        for (var i = 0; i < us.length; i++) {
+        Utilities.traversePairsObj(pairsObj, function (u, v) {
 
-            u = us[i];
-
-            var v;
-            var vs = Object.keys(edgeObj[u]);
-            for (var j = 0; j < vs.length; j++) {
-                v = vs[j];
-                normalized[u] = normalized[u] || {};
-                normalized[u][v] = matchWeights ? edgeObj[u][v].weight : true;
-            }
-        }
+            normalized[u] = normalized[u] || {};
+            normalized[u][v] = matchWeights ? pairsObj[u][v].weight : true;
+        });
 
         return normalized;
     };
@@ -435,10 +424,10 @@ internals.Graph.prototype.snapshot = function () {
     var pairsObj = this._edges;
 
     if (!this.directed) {
-        pairsObj = Formatting.halvePairsObj(pairsObj);
+        pairsObj = Utilities.halvePairsObj(pairsObj);
     }
 
-    var pairs = Formatting.pairsObjToArray(pairsObj);
+    var pairs = Utilities.pairsObjToArray(pairsObj);
 
     return this.subgraph({
         vertices: Object.keys(this._vertices),
@@ -485,10 +474,10 @@ internals.Graph.prototype.complement = function () {
     }
 
     if (!this.directed) {
-        edges = Formatting.halvePairsObj(edges);
+        edges = Utilities.halvePairsObj(edges);
     }
 
-    edges = Formatting.pairsObjToArray(edges);
+    edges = Utilities.pairsObjToArray(edges);
 
     definition.vertices = vertices;
     definition.edges = edges;
@@ -569,24 +558,15 @@ internals.Graph.prototype.union = function (graph) {
     var edgeInfo = Hoek.merge(Hoek.clone(graph._edges), this._edges);
 
     // For traversing
-    var edgeObj = this.directed ? edgeInfo : Formatting.halvePairsObj(edgeInfo);
+    var pairsObj = this.directed ? edgeInfo : Utilities.halvePairsObj(edgeInfo);
 
-    var u;
-    var us = Object.keys(edgeObj);
-    for (var j = 0; j < us.length; j++) {
+    Utilities.traversePairsObj(pairsObj, function (u, w) {
 
-        u = us[j];
-
-        var w;
-        var ws = Object.keys(edgeObj[u]);
-        for (var k = 0; k < ws.length; k++) {
-            w = ws[k];
-            // Fix destruction of pairs from the edge merge
-            edgeInfo[u][w].pair = [u, w];
-            edgeInfo[u][w].labels = Object.keys(edgeInfo[u][w].labels);
-            edges.push(edgeInfo[u][w]);
-        }
-    }
+        // Fix destruction of pairs from the edge merge
+        edgeInfo[u][w].pair = [u, w];
+        edgeInfo[u][w].labels = Object.keys(edgeInfo[u][w].labels);
+        edges.push(edgeInfo[u][w]);
+    });
 
     definition.vertices = vertices;
     definition.edges = edges;
@@ -623,27 +603,18 @@ internals.Graph.prototype.intersection = function (graph) {
     var theirEdgeInfo = graph._edges;
 
     // For traversing
-    var edgeObj = this.directed ? ourEdgeInfo : Formatting.halvePairsObj(ourEdgeInfo);
+    var pairsObj = this.directed ? ourEdgeInfo : Utilities.halvePairsObj(ourEdgeInfo);
 
-    var u;
-    var us = Object.keys(edgeObj);
-    for (var j = 0; j < us.length; j++) {
+    Utilities.traversePairsObj(pairsObj, function (u, w) {
 
-        u = us[j];
-
-        var w;
-        var ws = Object.keys(edgeObj[u]);
-        for (var k = 0; k < ws.length; k++) {
-            w = ws[k];
-            if (graph._lookupEdge(u, w)) {
-                edges.push({
-                    pair: [u, w],
-                    labels: Hoek.intersect(Object.keys(ourEdgeInfo[u][w].labels), Object.keys(theirEdgeInfo[u][w].labels)),
-                    weight: ourEdgeInfo[u][w].weight
-                });
-            }
+        if (graph._lookupEdge(u, w)) {
+            edges.push({
+                pair: [u, w],
+                labels: Hoek.intersect(Object.keys(ourEdgeInfo[u][w].labels), Object.keys(theirEdgeInfo[u][w].labels)),
+                weight: ourEdgeInfo[u][w].weight
+            });
         }
-    }
+    });
 
     definition.vertices = vertices;
     definition.edges = edges;
@@ -897,10 +868,10 @@ internals.Graph.prototype._getEdgePairsByLabel = function (label) {
     var pairsObj = label ? (this._edgeLabels[label] || {}) : this._edges;
 
     if (!this.directed) {
-        pairsObj = Formatting.halvePairsObj(pairsObj);
+        pairsObj = Utilities.halvePairsObj(pairsObj);
     }
 
-    return Formatting.pairsObjToArray(pairsObj);
+    return Utilities.pairsObjToArray(pairsObj);
 };
 
 internals.Graph.prototype._publicizeVertex  = function (vertex) {
