+++ lib/schema.js
@@ -48,4 +48,4 @@ exports.definition = Joi.object({
             weight: Joi.number()
         })
     ])
-});
+}).default();
