+++ lib/traversal.js
@@ -1,5 +1,5 @@
 var Hoek = require('hoek');
-var Formatting = require('./formatting');
+var Utilities = require('./utilities');
 
 var internals = {};
 
@@ -60,7 +60,7 @@ internals.Traversal.prototype.subgraph = function () {
 
     var subset = {
         vertices: Object.keys(this._visited),
-        edges: Formatting.pairsObjToArray(this._edges)
+        edges: Utilities.pairsObjToArray(this._edges)
     };
 
     return this.graph.subgraph(subset);
